pyproject.toml
[project]
name = "garmin-tracker"
version = "0.1.0"
description = "AI-powered Garmin health data tracker and analyzer"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn==0.34.0",
    "garminconnect==0.2.25",
    "anthropic==0.43.0",
    "jinja2==3.1.5",
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.8",
    "mypy>=1.13",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["garmin_tracker*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
